Fixes for header files, floating point trig functions, and incorrect conversions.

+++ flight/Libraries/WorldMagModel.c
@@ -179,7 +179,7 @@ int WMM_Initialize()
 	Ellip->a = 6378.137;	// semi-major axis of the ellipsoid in km
 	Ellip->b = 6356.7523142;	// semi-minor axis of the ellipsoid in km
 	Ellip->fla = 1 / 298.257223563;	// flattening
-	Ellip->eps = sqrt(1 - (Ellip->b * Ellip->b) / (Ellip->a * Ellip->a));	// first eccentricity
+	Ellip->eps = sqrtf(1 - (Ellip->b * Ellip->b) / (Ellip->a * Ellip->a));	// first eccentricity
 	Ellip->epssq = (Ellip->eps * Ellip->eps);	// first eccentricity squared
 	Ellip->re = 6371.2;	// Earth's radius in km
 
@@ -435,8 +435,8 @@ int WMM_ComputeSphericalHarmonicVariables(WMMtype_CoordSpherical *CoordSpherical
 	float cos_lambda, sin_lambda;
 	uint16_t m, n;
 
-	cos_lambda = cos(CoordSpherical->lambda * DEG2RAD);
-	sin_lambda = sin(CoordSpherical->lambda * DEG2RAD);
+	cos_lambda = cosf(CoordSpherical->lambda * DEG2RAD);
+	sin_lambda = sinf(CoordSpherical->lambda * DEG2RAD);
 
 	/* for n = 0 ... model_order, compute (Radius of Earth / Spherica radius r)^(n+2)
 	   for n  1..nMax-1 (this is much faster than calling pow MAX_N+1 times).      */
@@ -482,7 +482,7 @@ int WMM_AssociatedLegendreFunction(WMMtype_CoordSpherical * CoordSpherical, uint
 
 	 */
 {
-	float sin_phi = sin(CoordSpherical->phig * DEG2RAD);	/* sin  (geocentric latitude) */
+	float sin_phi = sinf(CoordSpherical->phig * DEG2RAD);	/* sin  (geocentric latitude) */
 
 	if (nMax <= 16 || (1 - fabs(sin_phi)) < 1.0e-10)	/* If nMax is less tha 16 or at the poles */
 	{
@@ -569,7 +569,7 @@ int WMM_Summation(WMMtype_LegendreFunction * LegendreFunction,
 		}
 	}
 
-	cos_phi = cos(CoordSpherical->phig * DEG2RAD);
+	cos_phi = cosf(CoordSpherical->phig * DEG2RAD);
 	if (fabs(cos_phi) > 1.0e-10)
 	{
 		MagneticResults->By = MagneticResults->By / cos_phi;
@@ -649,7 +649,7 @@ int WMM_SecVarSummation(WMMtype_LegendreFunction * LegendreFunction,
 			    * LegendreFunction->dPcup[index];
 		}
 	}
-	cos_phi = cos(CoordSpherical->phig * DEG2RAD);
+	cos_phi = cosf(CoordSpherical->phig * DEG2RAD);
 	if (fabs(cos_phi) > 1.0e-10)
 	{
 		MagneticResults->By = MagneticResults->By / cos_phi;
@@ -700,8 +700,8 @@ int WMM_RotateMagneticVector(WMMtype_CoordSpherical * CoordSpherical,
 	float Psi = (CoordSpherical->phig - CoordGeodetic->phi) * DEG2RAD;
 
 	/* Rotate spherical field components to the Geodeitic system */
-	MagneticResultsGeo->Bz = MagneticResultsSph->Bx * sin(Psi) + MagneticResultsSph->Bz * cos(Psi);
-	MagneticResultsGeo->Bx = MagneticResultsSph->Bx * cos(Psi) - MagneticResultsSph->Bz * sin(Psi);
+	MagneticResultsGeo->Bz = MagneticResultsSph->Bx * sinf(Psi) + MagneticResultsSph->Bz * cosf(Psi);
+	MagneticResultsGeo->Bx = MagneticResultsSph->Bx * cosf(Psi) - MagneticResultsSph->Bz * sinf(Psi);
 	MagneticResultsGeo->By = MagneticResultsSph->By;
 
     return 0;
@@ -729,10 +729,10 @@ int WMM_CalculateGeoMagneticElements(WMMtype_MagneticResults * MagneticResultsGe
 	GeoMagneticElements->Y = MagneticResultsGeo->By;
 	GeoMagneticElements->Z = MagneticResultsGeo->Bz;
 
-	GeoMagneticElements->H = sqrt(MagneticResultsGeo->Bx * MagneticResultsGeo->Bx + MagneticResultsGeo->By * MagneticResultsGeo->By);
-	GeoMagneticElements->F = sqrt(GeoMagneticElements->H * GeoMagneticElements->H + MagneticResultsGeo->Bz * MagneticResultsGeo->Bz);
-	GeoMagneticElements->Decl = atan2(GeoMagneticElements->Y, GeoMagneticElements->X) * RAD2DEG;
-	GeoMagneticElements->Incl = atan2(GeoMagneticElements->Z, GeoMagneticElements->H) * RAD2DEG;
+	GeoMagneticElements->H = sqrtf(MagneticResultsGeo->Bx * MagneticResultsGeo->Bx + MagneticResultsGeo->By * MagneticResultsGeo->By);
+	GeoMagneticElements->F = sqrtf(GeoMagneticElements->H * GeoMagneticElements->H + MagneticResultsGeo->Bz * MagneticResultsGeo->Bz);
+	GeoMagneticElements->Decl = atan2f(GeoMagneticElements->Y, GeoMagneticElements->X) * RAD2DEG;
+	GeoMagneticElements->Incl = atan2f(GeoMagneticElements->Z, GeoMagneticElements->H) * RAD2DEG;
 
     return 0;   // OK
 }
@@ -844,7 +844,7 @@ int WMM_PcupHigh(float *Pcup, float *dPcup, float x, uint16_t nMax)
 	scalef = 1.0e-280;
 
 	for (n = 0; n <= 2 * nMax + 1; ++n)
-		PreSqr[n] = sqrt((float)(n));
+		PreSqr[n] = sqrtf((float)(n));
 
 	k = 2;
 
@@ -863,7 +863,7 @@ int WMM_PcupHigh(float *Pcup, float *dPcup, float x, uint16_t nMax)
 	}
 
 	/*z = sin (geocentric latitude) */
-	z = sqrt((1.0 - x) * (1.0 + x));
+	z = sqrtf((1.0 - x) * (1.0 + x));
 	pm2 = 1.0;
 	Pcup[0] = 1.0;
 	dPcup[0] = 0.0;
@@ -978,7 +978,7 @@ int WMM_PcupLow(float *Pcup, float *dPcup, float x, uint16_t nMax)
 	dPcup[0] = 0.0;
 
 	/*sin (geocentric latitude) - sin_phi */
-	z = sqrt((1.0 - x) * (1.0 + x));
+	z = sqrtf((1.0 - x) * (1.0 + x));
 
 	/*       First, Compute the Gauss-normalized associated Legendre  functions */
 	for (n = 1; n <= nMax; n++)
@@ -1035,7 +1035,7 @@ int WMM_PcupLow(float *Pcup, float *dPcup, float x, uint16_t nMax)
 		{
 			index = (n * (n + 1) / 2 + m);
 			index1 = (n * (n + 1) / 2 + m - 1);
-			schmidtQuasiNorm[index] = schmidtQuasiNorm[index1] * sqrt((float)((n - m + 1) * (m == 1 ? 2 : 1)) / (float)(n + m));
+			schmidtQuasiNorm[index] = schmidtQuasiNorm[index1] * sqrtf((float)((n - m + 1) * (m == 1 ? 2 : 1)) / (float)(n + m));
 		}
 
 	}
@@ -1088,7 +1088,7 @@ int WMM_SummationSpecial(WMMtype_SphericalHarmonicVariables *
 	schmidtQuasiNorm1 = 1.0;
 
 	MagneticResults->By = 0.0;
-	sin_phi = sin(CoordSpherical->phig * DEG2RAD);
+	sin_phi = sinf(CoordSpherical->phig * DEG2RAD);
 
 	for (n = 1; n <= MagneticModel->nMax; n++)
 	{
@@ -1099,7 +1099,7 @@ int WMM_SummationSpecial(WMMtype_SphericalHarmonicVariables *
 
 		index = (n * (n + 1) / 2 + 1);
 		schmidtQuasiNorm2 = schmidtQuasiNorm1 * (float)(2 * n - 1) / (float)n;
-		schmidtQuasiNorm3 = schmidtQuasiNorm2 * sqrt((float)(n * 2) / (float)(n + 1));
+		schmidtQuasiNorm3 = schmidtQuasiNorm2 * sqrtf((float)(n * 2) / (float)(n + 1));
 		schmidtQuasiNorm1 = schmidtQuasiNorm2;
 		if (n == 1)
 		{
@@ -1154,13 +1154,13 @@ int WMM_SecVarSummationSpecial(WMMtype_SphericalHarmonicVariables *
 	schmidtQuasiNorm1 = 1.0;
 
 	MagneticResults->By = 0.0;
-	sin_phi = sin(CoordSpherical->phig * DEG2RAD);
+	sin_phi = sinf(CoordSpherical->phig * DEG2RAD);
 
 	for (n = 1; n <= MagneticModel->nMaxSecVar; n++)
 	{
 		index = (n * (n + 1) / 2 + 1);
 		schmidtQuasiNorm2 = schmidtQuasiNorm1 * (float)(2 * n - 1) / (float)n;
-		schmidtQuasiNorm3 = schmidtQuasiNorm2 * sqrt((float)(n * 2) / (float)(n + 1));
+		schmidtQuasiNorm3 = schmidtQuasiNorm2 * sqrtf((float)(n * 2) / (float)(n + 1));
 		schmidtQuasiNorm1 = schmidtQuasiNorm2;
 		if (n == 1)
 		{
@@ -1306,11 +1306,11 @@ int WMM_GeodeticToSpherical(WMMtype_CoordGeodetic * CoordGeodetic, WMMtype_Coord
 {
 	float CosLat, SinLat, rc, xp, zp;	// all local variables
 
-	CosLat = cos(CoordGeodetic->phi * DEG2RAD);
-	SinLat = sin(CoordGeodetic->phi * DEG2RAD);
+	CosLat = cosf(CoordGeodetic->phi * DEG2RAD);
+	SinLat = sinf(CoordGeodetic->phi * DEG2RAD);
 
 	// compute the local radius of curvature on the WGS-84 reference ellipsoid
-	rc = Ellip->a / sqrt(1.0 - Ellip->epssq * SinLat * SinLat);
+	rc = Ellip->a / sqrtf(1.0 - Ellip->epssq * SinLat * SinLat);
 
 	// compute ECEF Cartesian coordinates of specified point (for longitude=0)
 
@@ -1319,8 +1319,8 @@ int WMM_GeodeticToSpherical(WMMtype_CoordGeodetic * CoordGeodetic, WMMtype_Coord
 
 	// compute spherical radius and angle lambda and phi of specified point
 
-	CoordSpherical->r = sqrt(xp * xp + zp * zp);
-	CoordSpherical->phig = asin(zp / CoordSpherical->r) * RAD2DEG;	// geocentric latitude
+	CoordSpherical->r = sqrtf(xp * xp + zp * zp);
+	CoordSpherical->phig = asinf(zp / CoordSpherical->r) * RAD2DEG;	// geocentric latitude
 	CoordSpherical->lambda = CoordGeodetic->lambda;	// longitude
 
 	return 0;   // OK
